Move Subscriptions compatibility to its own class. (#3256)
* Start of compatibility work with Bookings and Multi-Currency.

* Added tests for Bookings and Multi-Currency compatibility.

* Add missing test file, update changelog and readme.

* Fixed display price, calculated price, and conversion in admin.

* Refactor how compatibility classes are initialized, finish tests.

* Move compatibility for Product Add Ons to its own class.

* Move Subscriptions compatibility to its own class.

* Fix FILTER_PREFIX constant.
